+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 * Added bigger image columns to element index
 * Added `Story` entry type for the `News` section.
 * Improved display of copyright notices in footer.
-* Added `requestData` app component and `request` twig variable, representing a Laravel collection that can be used to simulate a request scope (experimental).
+* Added `_globals` twig variable, representing a Laravel collection that can be used to simulate a request scope (experimental).
+* Added `customConfig` twig variable as alias for `craft.app.config.custom`.
 * Prepared for rtl languages (needs review).
 * Dropped environment variable fields from `Site Info` single entry field layout, as they do not support drafts. The fields are still available if you want to use them in a global set.
 * Added `HeaderWithBackgroundImage` hero area template